fix parentId issue
