+++ src/features/Conversation/Messages/Group/GroupChildren.tsx
@@ -1,12 +1,10 @@
 import { AssistantContentBlock } from '@lobechat/types';
 import { createStyles } from 'antd-style';
-import { memo, use } from 'react';
+import isEqual from 'fast-deep-equal';
+import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
-import { VirtuosoContext } from '@/features/Conversation/components/VirtualizedList/VirtuosoContext';
-import { useChatStore } from '@/store/chat';
-
-import { ContentBlock } from './ContentBlock';
+import GroupItem from './GroupItem';
 
 const useStyles = createStyles(({ css }) => {
   return {
@@ -30,35 +28,24 @@ const GroupChildren = memo<GroupChildrenProps>(
   ({ blocks, contentId, disableEditing, messageIndex, id }) => {
     const { styles } = useStyles();
 
-    const [toggleMessageEditing] = useChatStore((s) => [s.toggleMessageEditing]);
-    const virtuosoRef = use(VirtuosoContext);
-
     return (
       <Flexbox className={styles.container} gap={8}>
         {blocks.map((item, index) => {
-          return item.id === contentId ? (
-            <Flexbox
-              key={index}
-              onDoubleClick={(e) => {
-                if (disableEditing || item.error || !e.altKey) return;
-
-                toggleMessageEditing(item.id, true);
-                virtuosoRef?.current?.scrollIntoView({
-                  align: 'start',
-                  behavior: 'auto',
-                  index: messageIndex,
-                });
-              }}
-            >
-              <ContentBlock index={index} {...item} />
-            </Flexbox>
-          ) : (
-            <ContentBlock index={index} key={`${id}_${index}`} {...item} />
+          return (
+            <GroupItem
+              {...item}
+              contentId={contentId}
+              disableEditing={disableEditing}
+              index={index}
+              key={`${id}_${index}`}
+              messageIndex={messageIndex}
+            />
           );
         })}
       </Flexbox>
     );
   },
+  isEqual,
 );
 
 export default GroupChildren;
